refactor(admin): merge member edit and people pages
